requirements.txt
google-api-python-client==2.184.0
isodate>=0.6.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
requests>=2.31.0
pytest>=7.0.0
pytest-cov>=4.0.0
soda-core==3.5.5
soda-core-postgres==3.5.5
apache-airflow-providers-postgres>=5.0.0
setuptools>=65